requirements.txt
# Core dependencies for HootScape
python-dotenv
pyaudio
aiohttp
pydantic
pydantic-settings

# HootScape’s voice integration and pipeline services, with extras for OpenAI and Deepgram support.
pipecat-ai[openai,deepgram]

# Spatial audio processing for the immersive soundscape.
# spatialscaper

pyserial
gTTS
pydub